fix: transition to o200k_base vocab and resolve CUDA OOMs
- Migrated dataset pipeline from uint16/gpt2 to uint32/o200k_base to fix garbage text generation.
- Adjusted config batch_size (3) and grad accumulation (16) to fit larger vocab logits in 4GB VRAM.
- Disabled pin_memory in DataLoaders to prevent asynchronous CUDA driver faults on evaluation.
- Bypassed torch.compile during evaluation generation and checkpoint saving to ensure model stability and clean state dicts.

+++ src/tiny_shakespeare_gpt/dataset.py
@@ -11,7 +11,7 @@ class MemmapTokenDataset(Dataset):
     Using memory-mapping ensures that huge datasets load instantly and consume 
     almost zero RAM.
     """
-    def __init__(self, file_path: str, block_size: int, dtype: np.dtype = np.uint16):
+    def __init__(self, file_path: str, block_size: int, dtype: np.dtype = np.uint32):
         self.data = np.memmap(file_path, dtype=dtype, mode='r')
         self.block_size = block_size
